started Parsing from vector

+++ include/Webserv.hpp
@@ -36,12 +36,12 @@ class Server
         ~Server();
 
         static int epollInit(ServerList &servers);
-        // int run(FileDescriptor& fds);
+        // int run(FileDescriptor &fds);
         static int make_socket_non_blocking(int sfd);
-        // static int runServers(std::vector<Server>& servers, FileDescriptor& fds);
-        static int runServers(ServerList& servers, FileDescriptor& fds);
-        static void handleEvents(ServerList& servers, FileDescriptor& fds, int eventCount);
-        static void acceptConnection(const std::unique_ptr<Server> &server, FileDescriptor& fds);
+        // static int runServers(std::vector<Server> &servers, FileDescriptor &fds);
+        static int runServers(ServerList &servers, FileDescriptor &fds);
+        static void handleEvents(ServerList &servers, FileDescriptor &fds, size_t eventCount);
+        static void acceptConnection(const std::unique_ptr<Server> &server, FileDescriptor &fds);
     private:
         std::string _serverName;
         int _listener;
